Improve HTML escaping
Borrowed heavily from @andersk in #54

I didn't feel the complexity of having a title and title_html was merited.
The soft-hyphen was originally added to improve padding
on the landing page, but "Reinforcement Learning"
had the same issue as "Metaheuristics" so there isn't a downside
to removing the soft-hyphen completely now, and longer-term work toward
a design that is more accomodating to long topic names.